+++ packages/terra-core-docs/src/terra-dev-site/test/profile-image/ProfileImageDefault.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import ProfileImage from '../../../ProfileImage';
-import exampleProfileImage from '../../assets/150x150.jpg';
+import ProfileImage from 'terra-profile-image';
+import exampleProfileImage from '../../doc/profile-image/assets/150x150.jpg';
 
 const ProfileImageDefault = () => (
   <div>